requirements.txt
requests
kafka-python
spark 
pyspark
cassandra-driver